move s3 config to separate file
